# Reuse Chunk Iterator