--- a/querier.go
+++ b/querier.go
@@ -1060,9 +1060,10 @@ func (it *verticalMergeSeriesIterator) Err() error {
 type chunkSeriesIterator struct {
 	chunks []chunks.Meta
 
-	i       int
-	cur     chunkenc.Iterator
-	bufIter *chunkenc.XORIterator
+	i          int
+	cur        chunkenc.Iterator
+	bufXorIter *chunkenc.XORIterator
+	bufDelIter *deletedIterator
 
 	maxt, mint int64
 
@@ -1085,18 +1086,24 @@ func newChunkSeriesIterator(cs []chunks.Meta, dranges Intervals, mint, maxt int6
 }
 
 func (it *chunkSeriesIterator) resetCurIterator() {
-	if it.bufIter != nil {
-		it.bufIter.Reset(it.chunks[it.i].Chunk.Bytes())
-		it.cur = it.bufIter
+	if it.bufXorIter != nil {
+		it.bufXorIter.Reset(it.chunks[it.i].Chunk.Bytes())
+		it.cur = it.bufXorIter
 	} else {
 		it.cur = it.chunks[it.i].Chunk.Iterator()
 		if xorIter, ok := it.cur.(*chunkenc.XORIterator); ok {
-			it.bufIter = xorIter
+			it.bufXorIter = xorIter
 		}
 	}
-	if len(it.intervals) > 0 {
-		it.cur = &deletedIterator{it: it.cur, intervals: it.intervals}
+	if len(it.intervals) == 0 {
+		return
 	}
+	if it.bufDelIter == nil {
+		it.bufDelIter = &deletedIterator{}
+	}
+	it.bufDelIter.it = it.cur
+	it.bufDelIter.intervals = it.intervals
+	it.cur = it.bufDelIter
 }
 
 func (it *chunkSeriesIterator) Seek(t int64) (ok bool) {