Want feedback if something like this would be acceptable. 

When there are millions of chunks to be parsed for a query, re-using the chunk iterator saves a lot of allocs and in my benchmark in cortex it shed up to `5+s` for a query. (I don't have an exact number of chunks as of now).