@codesome 2019-06-28
Pointing out the function which causes the allocs
https://github.com/prometheus/tsdb/blob/ee010e9fab65565d65d851ac0b5b908c3e215798/chunkenc/xor.go#L105-L113