@codesome 2019-06-28
> Reset should always work.

For `deletedIterator`, how would `Reset` work? Because it needs deleted intervals to be reset, but the argument takes only `[]byte`. If we generalise it to `interface{}`, it would end up taking more allocs.