@brian-brazil 2019-06-28
Sounds like this isn't quite the right abstraction then. How about something more type specific?