@codesome 2019-06-28
You mean to say have that method to the struct and deference the interface to the type in order to call that method when needed?