@brian-brazil 2019-06-28
Yes, it really depends on what the call site is.