@brian-brazil 2019-06-28
I would like to see what the benchmarks look like for this, as when I tested things like this previously it didn't help much.

Does this change any of the API semantics?