@codesome 2019-06-28
Tested on the same query and data as described in [this comment](https://github.com/prometheus/prometheus/pull/5707#issuecomment-505440230). Also the code changes that I made in cortex querier to benchmark this is exactly similar to what I did in `querier.go` in this PR.
```
benchmark                                                                                                                    old ns/op       new ns/op       delta
BenchmarkChunkQueryableFromTar/query=count(metric_name{label1="value1"}),from=1561346880,to=1561368480,step=15.000000-8     41026506987     39711348531     -3.21%

benchmark                                                                                                                    old allocs     new allocs     delta
BenchmarkChunkQueryableFromTar/query=count(metric_name{label1="value1"}),from=1561346880,to=1561368480,step=15.000000-8     6479152        3027294        -53.28%

benchmark                                                                                                                    old bytes      new bytes      delta
BenchmarkChunkQueryableFromTar/query=count(metric_name{label1="value1"}),from=1561346880,to=1561368480,step=15.000000-8     5500794632     5176138576     -5.90%
```

With repeated benchmarks, apparently, the gain in time seems to decrease. So it's mostly allocs and B.op.