@codesome 2019-06-28
> Does this change any of the API semantics?

I am not actually sure what you mean here.