@brian-brazil 2019-06-28
Currently a caller of this code can keep multiple of the returned iterators around, now it's only valid until the next Seek/Next.