@codesome 2019-07-03
> now it's only valid until the next Seek/Next.

I am setting the iterators while the iterator is being created https://github.com/prometheus/tsdb/blob/cef83448cdf1fd61850b870976786dddcd99dee7/querier.go#L1082
So I think it's valid as soon as it is created (given that it's created using `newChunkSeriesIterator`)