@codesome 2019-07-04
Now I am also re-using `deletedIterator`. What part about this doesn't feel right @brian-brazil ?