@codesome 2019-07-04
I have come up with another way to re-use the iterator. `chunckenc.Chunk.Iterator` method now takes a `chunckenc.Iterator` interface as an argument for reuse.