@codesome 2019-07-05
(Will hold off merging until 2.11.0 is out, just in case if there is a bugfix to be included from tsdb in the release)